Move to URI style backend
This change moves from a one-time backend to a URI based
hiera backend wherein the lookup_key function will be called for
each configured URI.

This change moves the paths settings out of the options hash to
the URIs array and uses the mounts options to control what kind of
secret backend is being used. It has been tested against kv2, and
assumed to work with kv1, structures are in place for other path based
backends if needed.

This change requires a modified version of the vault ruby gem which
correctly calls LIST method endpoints to help reduce the number of calls
being made to vault during the lookup process.

This change introduces the `environment_delimeter` option which will
allow the fields at a secret path to be selected based on the current
environment, bare fields will be selected by default. This feature
behaves as expected when default field is used.

This change will require updating the configuration to use uris
instead of putting the paths in the options block.