Add retry and Drop gem PR Requirement
This change resolves a race condition, especially with cached environments,
where the client connection pool will shutdown during a requst resulting
in a catalog compilation failure by using the built in retry
functionality of the vault-gem.

This change also moves to use the `Vault::logical.list` method over
requiring the `LIST` HTTP method as part of PR 201 on the vault gem.
The workaround resulted from a bug in the `vault agent` request
forwarding whereby query string parameteres failed to be forwarded.
Using the latest `vault agent` allows the typical method to be used.