@jovandeginste's work on strict mode, now with tests 👍🏻  (#78)
* Bailout on http errors, add strict_mode

We let the lookup function fail in case of http errors when looking up a
secret. Furthermore, we add strict_mode, that if set, will fail puppet
when a secret cannot be found in any of the paths. This should only be
used in combination of confine_to_keys.

Signed-off-by: Peter Verraedt <peter.verraedt@kuleuven.be>
Signed-off-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>

* Strict_mode detection needs to be parenthesised

* Otherwise we can’t detect the edgecase of it being
set but set explicitly to false

* Adding Spec file

* Usecases to test:
Setting set: Don’t raise error
Setting set and false: Don’t raise error
Setting not set: Don’t raise error

* Changing error message

- Explictly mention strict mode
- Move msg into a variable

Co-authored-by: Jo Vandeginste <Jo.Vandeginste@kuleuven.be>